// src/types.ts
export interface NavItem {
  title: string;
  href: string;
}

export interface Feature {
  title: string;
  description: string;
  icon: string;
}

export interface CaseStudy {
  title: string;
  description: string;
  videoUrl: string;
  results: string[];
}

export interface Creator {
  name: string;
  image: string;
}